lawyer_agent(tingli)/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch>=1.13.0
transformers>=4.21.0
scikit-learn==1.3.0
pandas==2.0.3
flask==2.3.3
flask-cors==4.0.0
pydantic==2.4.2
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
pathlib